app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   flask
    #   pystow
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses==0.6
    # via useful-rdkit-utils
flasgger==0.9.7.1
    # via -r requirements.in
flask==3.1.1
    # via
    #   -r requirements.in
    #   flasgger
    #   flask-cors
flask-cors==6.0.1
    # via -r requirements.in
fonttools==4.59.1
    # via matplotlib
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.1
    # via scikit-learn
jsonschema==4.25.1
    # via flasgger
jsonschema-specifications==2025.4.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
loguru==0.7.3
    # via scaffoldgraph
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.5
    # via
    #   scaffoldgraph
    #   seaborn
mistune==3.1.3
    # via flasgger
networkx==3.5
    # via scaffoldgraph
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   rdkit
    #   scikit-learn
    #   scipy
    #   seaborn
    #   useful-rdkit-utils
packaging==25.0
    # via
    #   flasgger
    #   gunicorn
    #   matplotlib
    #   pytest
pandas==2.3.1
    # via
    #   -r requirements.in
    #   seaborn
    #   useful-rdkit-utils
pillow==11.3.0
    # via
    #   matplotlib
    #   rdkit
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements.in
py3dmol==2.5.2
    # via useful-rdkit-utils
pyarrow==21.0.0
    # via useful-rdkit-utils
pygments==2.19.2
    # via pytest
pyparsing==3.2.3
    # via matplotlib
pystow==0.7.6
    # via useful-rdkit-utils
pytest==8.4.1
    # via
    #   -r requirements.in
    #   scaffoldgraph
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via flasgger
rdkit==2024.3.3
    # via
    #   -r requirements.in
    #   useful-rdkit-utils
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via pystow
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
scaffoldgraph[rdkit]==1.1.1
    # via -r requirements.in
scikit-learn==1.7.1
    # via useful-rdkit-utils
scipy==1.16.1
    # via
    #   scaffoldgraph
    #   scikit-learn
    #   useful-rdkit-utils
seaborn==0.13.2
    # via useful-rdkit-utils
six==1.17.0
    # via
    #   flasgger
    #   python-dateutil
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   pystow
    #   scaffoldgraph
    #   useful-rdkit-utils
typing-extensions==4.14.1
    # via
    #   pystow
    #   referencing
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
useful-rdkit-utils==0.89
    # via -r requirements.in
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors

# The following packages are considered to be unsafe in a requirements file:
# setuptools
